+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-appmod/src/main/java/com/microsoft/azure/toolkit/intellij/appmod/javamigration/MigrateToAzureNode.java
@@ -92,7 +92,7 @@ public final class MigrateToAzureNode extends Node<String> {
         log.debug("[MigrateToAzureNode] showNotInstalled - copilotInstalled: {}", copilotInstalled);
         
         // Dynamic description based on what needs to be installed
-        final String description = "Install Github Copilot modernization";
+        final String description = "Install GitHub Copilot modernization";
         withDescription(description);
         
         onClicked(e -> {