+++ CHANGELOG.md
@@ -125,7 +125,7 @@ All notable changes to "Azure Toolkit for IntelliJ IDEA" will be documented in t
 
 ## 3.97.3
 ### Fixed
-- Rename Github Copilot app modernization to GitHub Copilot modernization.
+- Rename GitHub Copilot app modernization to GitHub Copilot modernization.
 
 ## 3.97.2
 ### Fixed