+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/resources/whatsnew.md
@@ -3,7 +3,7 @@
 
 ## 3.97.3
 ### Fixed
-- Rename Github Copilot app modernization to GitHub Copilot modernization.
+- Rename GitHub Copilot app modernization to GitHub Copilot modernization.
 
 ## 3.97.2
 ### Fixed